requirements.txt
# Core data science libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Flask and web framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Data formats
pyarrow>=12.0.0  # For parquet support
openpyxl>=3.1.0  # For Excel support

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0

# Deep learning
tensorflow>=2.15.0
keras>=2.15.0

# Time series analysis
statsmodels>=0.14.0
prophet>=1.1.5

# Advanced ML
xgboost>=2.0.0
lightgbm>=4.1.0

# Model optimization
optuna>=3.5.0

# AI Integration
google-generativeai>=0.3.0
